requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
